Factor out statements.py
Moved routines from typemap.py.  Statements and typemap have separate
functionality.  Statements started in the same file before it was
fully developed.  Time to move out.
