+++ shroud/ast.py
@@ -14,6 +14,7 @@ import copy
 
 from . import util
 from . import declast
+from . import statements
 from . import todict
 from . import typemap
 from .util import wformat
@@ -378,7 +379,7 @@ class LibraryNode(AstNode, NamespaceMixin):
         # Create typemaps once.
         if not typemap.get_global_types():
             typemap.initialize()
-        typemap.update_typemap_for_language(self.language)
+        statements.update_typemap_for_language(self.language)
 
         self.setup = kwargs.get("setup", {}) # for setup.py
